Fix clipping when colliding with two surfaces
To do this we simply check collision on each side of the player with
any collision object. This is very bad but can be optimized in a simple
way down the line. (think quadtrees or even simpler just check for
objects in the view)

Also adjusted the zoom level a bit!