Add tentative InputManager and pinch to zoom feature
This commit adds a barebones input manager class and the pinch to zoom
feature. This can be tested using the android emulator by holding the
ctrl/cmd key and dragging the mouse around.

I want to try and make the InputManager an actual manager but I'm not
sure of its design yet, hence 'tentative'.