Launch fight on collision with enemy
The player colliding with the enemy's fight rectangle launches a fight screen
(delay of 2 seconds while waiting for transition screens). A screenshot of the
enemy has been added to the fight screen. After the fight, the player is
repositioned away from the enemy so it doesn't trigger a new fight immediately.