Fix exploration UI scaling (sort of)
So for now I've settled on using letterboxing for non-16:9 aspect
ratios.
This means that the game will be scaled to the screen, no matter which
screen, but for unsupported aspect ratios the remaining space will be
filled up by a black bar.
This way we still achieve our goal of resolution independence without
sacrificing simplicity of rendering.