Add initial implementation of action zone
When the player gets into the 'action zone' of an entity, some kind of
icon will appear above that entity, indication that the player can
interact with said entity. Currently there is a question mark icon that
floats above an entity when the player is inside of this interaction
radius. All the player has to do to interact, is tap on the entity.