Clean up entity loading and update ? icon/logic
Includes a bigger number of changes:

- Update map to include player/enemy spawn positions
- Update all entity classes to allow for this change
- Update ? icon and accompanying logic
- Probably something else I forgot

The main purpose was to clean up our main class. Now we're dependent on
the sizes defined in the map for each entity but it should really just
depend on the texture. However, we don't really have a fixed standard
for those so that is pretty much impossible (for now). The way it is now
should be more readable and more manageable than previously. Just a hop
in a good direction :shrug: