:sparkles: deprecate need to install fonts
MGCB is capable of discovering fonts for `spritefont`s if the font, with
the exact font name as described in the `spritefont` (i.e.: "Pixel
Operator" -> "Pixel Operator.ttf"), is present in the same directory as
the `spritefont` itself.

This deprecates the need for installing fonts locally just to build and
run the project.

Closes #59